requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
PyMySQL>=1.0.0
cryptography>=38.0.0
Werkzeug>=2.0.0
itsdangerous>=2.0.0
Jinja2>=3.0.0
pytz>=2022.0
pymysql>=1.0.0
matplotlib>=3.5.0
gunicorn>=20.1.0
Flask-Sessionstore>=0.4.5
Flask-Session>=0.4.0
oauthlib>=3.2.0
requests-oauthlib>=1.3.0 